requirements.txt
chainlit
python-dotenv
azure-ai-projects
azure-identity
azure-storage-blob
aiohttp
azure-ai-agents